Use proper Python packaging
