+++ Procfile
@@ -1,3 +1,3 @@
-web: NEW_RELIC_CONFIG_FILE=newrelic.ini newrelic-admin run-program gunicorn main:app
+web: NEW_RELIC_CONFIG_FILE=newrelic.ini newrelic-admin run-program gunicorn anyway:app
 worker: celery worker -A clusters_calculator -D
