+++ anyway/__init__.py
@@ -1,0 +1,1 @@
+from .flask_app import app
