+++ anyway/constants.py
