+++ anyway/database.py
@@ -1,7 +1,7 @@
 from sqlalchemy import create_engine
 from sqlalchemy.orm import scoped_session, sessionmaker
 from sqlalchemy.ext.declarative import declarative_base
-import config
+from . import config
 
 
 engine = create_engine(config.SQLALCHEMY_DATABASE_URI, convert_unicode=True, echo=False)
