+++ requirements.txt
@@ -1,6 +1,6 @@
 Flask==0.10.1
-Flask-Assets==0.10
-Flask-Babel==0.9
+Flask-Assets==0.12
+Flask-Babel==0.11.2
 Flask-SQLAlchemy==2.1
 Flask-Login==0.3.2
 Jinja2==2.7.1
@@ -27,3 +27,4 @@ apscheduler==2.1.2
 Flask-Compress==1.3.0
 rauth==0.7.2
 requests==2.4.3
+click
